add RegexExtractedValueTest
